create popup-controls-component and replace all methods to it
